requirements.txt
dataclasses==0.6 # For Python 3.6.
python-telegram-bot==12.5.1
